Βελτιώσεις ασφαλείας Σε κάθε φόρτωμα playlist δεν ξεκινάει αυτόματα νέο τραγούδι Δεν ξεκινάει αυτόματα να παίζει τραγούδι στο αρχικό φόρτωμα της εφαρμογής Διορθώσεις

+++ AJAX/registerUser.php
@@ -4,16 +4,18 @@
  * Created by rocean
  * Date: 08/06/16
  * Time: 17:35
+ * 
  * Εισαγωγή αρχικού χρήστη
+ * 
  */
 
 
 
 require_once('../libraries/common.inc.php');
 
-Page::checkValidAjaxRequest();
-
 session_start();
+
+Page::checkValidAjaxRequest(false);
 
 $conn = new RoceanDB();
 $lang = new Language();
@@ -27,27 +29,31 @@ if(isset($_GET['password']))
 if (isset($_GET['email']))
     $email=ClearString($_GET['email']);
 
+// Ελέγχει αν υπάρχει admin χρήστης ήδη.
+if(!$conn->CheckIfThereIsAdminUser()) {
+    $register = $conn->CreateUser($username, $email, $password, '1', 'local', null, null);
+
+    if ($register['success']) {
+        $jsonArray = array('success' => true);
+        RoceanDB::insertLog('User ' . $username . ' registered'); // Προσθήκη της κίνησης στα logs
+    } else {
+        $jsonArray = array('success' => false);
+    }
 
 
-$register=$conn->CreateUser($username, $email, $password, '1', 'local', null, null);
 
-if($register['success']) {
-    $jsonArray=array( 'success'=>true);
-    RoceanDB::insertLog('User '.$username.' registered'); // Προσθήκη της κίνησης στα logs
+    // ελέγχει και εισάγει τις αρχικές τιμές στον πίνακα options
+    OWMP::startBasicOptions();
+
+    // Δημιουργεί event που σβήνει logs που είναι παλιότερα των 30 ημερών και τρέχει κάθε μέρα
+    $eventQuery='DELETE FROM logs WHERE log_date<DATE_SUB(NOW(), INTERVAL 30 DAY)';
+    RoceanDB::createMySQLEvent('logsManage', $eventQuery, '1 DAY');
+
+    //Page::createCrontab(); // Προσθέτει τον demon στο crontab
+
+} else {
+    $jsonArray = array('success' => false);
 }
-else {
-    $jsonArray=array( 'success'=>false);
-}
-
-// ελέγχει και εισάγει τις αρχικές τιμές στον πίνακα options
-OWMP::startBasicOptions();
-
-// Δημιουργεί event που σβήνει logs που είναι παλιότερα των 30 ημερών και τρέχει κάθε μέρα
-$eventQuery='DELETE FROM logs WHERE log_date<DATE_SUB(NOW(), INTERVAL 30 DAY)';
-RoceanDB::createMySQLEvent('logsManage', $eventQuery, '1 DAY');
-
-//Page::createCrontab(); // Προσθέτει τον demon στο crontab
-
 
 
 echo json_encode($jsonArray, JSON_UNESCAPED_UNICODE);
